backend/requirements.txt
Django~=3.1.4
djangorestframework==3.11.0
django-rest-auth==0.9.5
django-allauth==0.40.0
Pillow==8.2.0